+++ man-roxygen/api_type.R
@@ -1,2 +1,2 @@
-#' @param api_type character; one of "REST", "SOAP", or "Bulk" indicating 
-#' which API to use when making the request
+#' @param api_type character; one of "REST", "SOAP", "Bulk 1.0", "Bulk 2.0", or 
+#' "Chatter" indicating which API to use when making the request