+++ NEWS.md
@@ -7,6 +7,7 @@
   * Allow users to control whether bulk query results are kept as all character or 
   the types are guessed (#12)
   * Add `sf_get_all_jobs_bulk()` so that users can see retrieve details for all bulk jobs (#13)
+  * Add new utility functions `sf_set_password()` and `sf_reset_password()` (#11)
   
 ### Bug Fixes